pyproject.toml
[tool.poetry]
name = "parserkiosk"
version = "0.1.0"
description = "Test your bloody parsers"
authors = ["aarnav <aarnavbos@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
python-box = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
python-lsp-server = "^1.3.3"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
include = '\.pyi?$'
extend-include = true
exclude = '''
/(
    \.git
  | \.toml
  | \.lock
  | \.md
  | \.env
)/
'''

[tool.isort]
profile = "black"
